refactor alphafoldd job code
