--- src/controllers/jobsController.ts
@@ -1,1 +1,0 @@
-export * from './jobs/index.js'
